리뷰이후/fix(YieldCalculator) : 수익률에 대한 결과가 int를 초과할 수 있으므로 long으로 받는다.

+++ src/main/java/domain/YieldCalculator.java
@@ -3,13 +3,13 @@ package domain;
 public class YieldCalculator {
 
     public static double calculateYield(final WinningResult winningResult, final Amount purchaseAmount) {
-        int totalProfit = calculateProfit(winningResult);
+        long totalProfit = calculateProfit(winningResult);
         return (double) totalProfit / purchaseAmount.getValue();
     }
 
-    private static int calculateProfit(final WinningResult winningResult) {
+    private static long calculateProfit(final WinningResult winningResult) {
         return winningResult.getValue().entrySet().stream()
-                .mapToInt(entry -> entry.getKey().getPrize() * entry.getValue())
+                .mapToLong(entry -> (long) entry.getKey().getPrize() * entry.getValue())
                 .sum();
     }
